* refactoring

+++ trunk/iGaia/Shaders/ShaderWater.frag
@@ -1,5 +1,5 @@
 const char* ShaderWaterF = STRINGIFY(
-                                                   varying mediump vec4   OUT_TexCoord;
+                                                   varying highp vec4   OUT_TexCoord;
                                                    uniform sampler2D EXT_TEXTURE_01;
                                                    uniform sampler2D EXT_TEXTURE_02;
                                                    uniform sampler2D EXT_TEXTURE_03;
@@ -7,7 +7,7 @@ const char* ShaderWaterF = STRINGIFY(
                                        
 void main(void)
 {
-    lowp vec2 vTexCoordProj = OUT_TexCoord.xy;
+    highp vec2 vTexCoordProj = OUT_TexCoord.xy;
 	vTexCoordProj.x = 0.5 - 0.5 * vTexCoordProj.x / OUT_TexCoord.w;
 	vTexCoordProj.y = 0.5 + 0.5 * vTexCoordProj.y / OUT_TexCoord.w;
 	vTexCoordProj = clamp(vTexCoordProj, 0.001, 0.999);
@@ -15,8 +15,8 @@ void main(void)
     
     vTexCoordProj = OUT_TexCoord.xy;
     vTexCoordProj.x = 0.5 + 0.5 * vTexCoordProj.x / OUT_TexCoord.w;
-	vTexCoordProj.y = 0.5 + 0.5 * vTexCoordProj.y / OUT_TexCoord.w;
-	vTexCoordProj = clamp(vTexCoordProj, 0.001, 0.999);
+    vTexCoordProj.y = 0.5 + 0.5 * vTexCoordProj.y / OUT_TexCoord.w;
+    vTexCoordProj = clamp(vTexCoordProj, 0.001, 0.999);
     lowp vec4 vRefractionColor = texture2D(EXT_TEXTURE_02, vTexCoordProj);
     gl_FragColor = mix(vReflectionColor, vRefractionColor, 0.6);
 }
